Major update to symlink script, add zlogin with link

+++ utilityScripts/symlinkAll.sh
@@ -1,21 +1,34 @@
-#!/bin/bash
+#!/bin/sh
 ###########################################################
-# This script creates symlinks for every $file in $dir to ~
+# This script creates symlinks for my dotfiles to ~
 ###########################################################
 
-## VARIABLES
+# Variables
 dir=/Users/yannvanhalewyn/dotfiles
-files="bash_profile bash_prompt gitignore_global vimrc gvimrc gitconfig profile tmux.conf functions aliases osx vim nvim nvimrc zshrc cheat"
+files="bash_profile bash_prompt gitignore_global
+"
 
-## GOT TO DIR
+# Got to dir
 echo "Changing to the $dir directory"
 cd $dir
-echo "...done"
 
-## DELETING .FILES AND CREATING SYMLINKS
+# Get files
+files=$(ls -A | grep -v 'DS_Store\|.git\|Scripts')
+echo "Will create symlinks for:"
+echo "$files"
+
+# Prompt user
+echo "Continue? (Y/n) \c"; read -n1 ans;
+if [ $ans != 'y' ]
+then
+  echo "\nOk! Some other time then."
+  exit
+fi
+
+# Deleting exising .links and creating symlinks
 for file in $files; do
-    echo "Deleting .$file"
-    rm ~/.$file
-    echo "Creating symlink for .$file"
+    echo "Deleting ~/.$file"
+    rm $HOME/.$file
+    echo "Linking $dir/$file -> ~/.$file"
     ln -s $dir/$file ~/.$file
 done
